refactor(Repository) extract Core_Read_Interface
